Compute reported products in the type they widen to
bugprone-implicit-widening-of-multiplication-result reports 455 products
computed in int (or unsigned int) and then converted to a wider type, such
as pulse and second counts passed as long event delays, sizes, and time_t
offsets. A cast to the destination type now leads the left operand, so the
product itself is computed wide. Where the product is spelled inside a
macro (PULSE_VIOLENCE, SECS_PER_MUD_DAY), the cast converts the macro's
value explicitly; for pointer offsets it leads the offset's first operand.
NODE_ADVANCE in the gameplay test adds the cast inside its body.

The edit was made by a script from the clang-tidy report (issue #218,
docs/ongoing-projects/clang-tidy-findings.md, step 4).

+++ src/character/class.c
@@ -3781,7 +3781,7 @@ long int level_exp(struct char_data *ch, int level)
     level--;
     if (level < 0)
       level = 0;
-    exp += (level * level * factor);
+    exp += ((long)level * level * factor);
     break;
 
   default:
